+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [v0.3.1] - 2025-10-20
+
+### Changed
+- Received event payload bodies will no longer be parsed if the event doesn't
+  define any hooks
+
+
 ## [v0.3.0] - 2025-09-19
 
 ### Added
@@ -65,3 +72,4 @@ Initial release 🎉
 [v0.1.1]: https://github.com/trag1c/monalisten/compare/v0.1.0...v0.1.1
 [v0.2.0]: https://github.com/trag1c/monalisten/compare/v0.1.1...v0.2.0
 [v0.3.0]: https://github.com/trag1c/monalisten/compare/v0.2.0...v0.3.0
+[v0.3.1]: https://github.com/trag1c/monalisten/compare/v0.3.0...v0.3.1